server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Frasier",email = "fsundra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-stubs (>=5.2.7,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
